backend/requirements.txt
flask
numpy
pandas
torch
transformers
scipy
pickle-mixin
nltk
SpeechRecognition
pyttsx3
pillow
matplotlib
